+++ src/Ds/Bundle/PortalApiBundle/DependencyInjection/Configuration.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Ds\Bundle\PortalApiBundle\DependencyInjection;
+
+use Symfony\Component\Config\Definition\Builder\TreeBuilder;
+use Symfony\Component\Config\Definition\ConfigurationInterface;
+use Oro\Bundle\ConfigBundle\DependencyInjection\SettingsBuilder;
+
+/**
+ * Class Configuration
+ */
+class Configuration implements ConfigurationInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getConfigTreeBuilder()
+    {
+        $treeBuilder = new TreeBuilder;
+        $rootNode = $treeBuilder->root('ds_portal_api');
+
+        return $treeBuilder;
+    }
+}