+++ src/Ds/Bundle/PortalApiBundle/DsPortalApiBundle.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Ds\Bundle\PortalApiBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+/**
+ * Class DsPortalApiBundle
+ */
+class DsPortalApiBundle extends Bundle
+{
+}