+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This changelog references the relevant changes (new features, changes and bugs) 
   * Updated account user persona page to use definition title.
   * Updated flatly-custom theme.
   * Updated copyright widget year.
+  * Added PortalApi bundle.
 
 
 # CHANGELOG for 0.3.0